pyproject.toml
[project]
name = "coinbasket"
version = "0.0.1"
description = "🧺 coinbasket"
authors = [
    {name = "mehdidotcafe",email = "contact@mehdi.cafe"}
]
license = {text = "GPL3"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "uagents (>=0.22.1,<0.23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ruff (>=0.11.4,<0.12.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
coinbasket = "coinbasket:main"